# api/openapi.yaml
openapi: 3.1.0
info:
  title: kafkito REST API
  version: "1"
  summary: Read/write access to Kafka clusters managed by kafkito.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  description: |
    JSON surface that backs the kafkito web UI. It is primarily a UI
    contract, but stable enough for scripting with `curl`.

    This document is the single source of truth for the HTTP contract
    (ADR-0005). The frontend TypeScript types are generated from it and a Go
    test asserts that every registered route is described here.

    Authentication: every `/api/v1/*` request is validated by kafkito's auth
    middleware. Depending on `KAFKITO_AUTH_MODE` it expects an
    `Authorization: Bearer <JWT>` header (`mock`, `oidc`, `xsuaa` in
    `-tags btp` builds) and answers 401 when it is missing or invalid, or injects a
    synthetic principal (`off`, `-tags devauth` builds only). kafkito has no
    login flow of its own:
    interactive login is handled by an upstream auth proxy (for example the
    SAP approuter or oauth2-proxy), which forwards the bearer token. The validated principal
    feeds the optional RBAC policy. `/healthz` and `/readyz` are never
    authenticated.

    Private (browser-stored) clusters: use the cluster path segment
    `__private__` and send the cluster definition base64-encoded in the
    `X-Kafkito-Cluster` header. Operators can disable private clusters or
    restrict them to a role with `private_clusters.mode`
...

    Conventions:
...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/info:
    get:
      operationId: getInfo
      tags: [meta]
      summary: Build info.
      responses:
        "200":
          description: OK
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/InfoResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /api/v1/openapi.yaml:
    get:
      operationId: getOpenApiSpec
      tags: [meta]
      summary: This OpenAPI document.
      responses:
        "200":
          description: The raw OpenAPI 3.1 document.
          headers:
            X-Request-Id: { $ref: "#/components/headers/RequestId" }
          content:
            application/yaml:
              schema: { type: string }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        JWT validated by kafkito (auth modes `mock`, `oidc` and `xsuaa`). In mode
        `off` (`-tags devauth` builds only) no token is required and a
        synthetic principal is used.
  headers:
    RequestId:
      description: >
        Request correlation id, set on every response. Reuses the inbound
        `X-Vcap-Request-Id`, the trace-id of a W3C `traceparent` or an inbound
        `X-Request-Id` when present (in that order), otherwise freshly
        generated. Matches the `request_id` field in the server logs.
      required: true
      schema: { type: string }
...
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token.
      headers:
        X-Request-Id: { $ref: "#/components/headers/RequestId" }
        WWW-Authenticate:
          description: Always `Bearer realm="kafkito"`.
          schema: { type: string }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      description: Error body emitted by all JSON error paths.
      properties:
        error:
          type: string
          description: Human-readable message (`unauthorized`, `forbidden`, `not found`, ...).
        code:
          type: string
          description: >
            Machine-readable code where one exists, e.g. `kafka_upstream`,
            `private_cluster_address_blocked`, `private_clusters_disabled`,
            `private_clusters_forbidden`,
            `production_confirmation_required`, `copy_concurrency_limit`,
            `kafka_message_too_large`, `kafka_not_authorized`, `rbac_denied`,
            `topic_consumers_timeout`, `value_masked`, `invalid_request` (the
            request does not match this document; `error` names the parameter
            or body field and the violated rule, never the submitted value).
        message:
          type: string
          description: Detail for 401 responses from the auth middleware.
        resource:
          type: string
          description: RBAC-denied resource as `<type>:<name>` (403 only).
        action:
          type: string
          description: RBAC-denied action (403 only).
...
    InfoResponse:
      type: object
      required: [name, version]
      properties:
        name: { type: string }
        version: { type: string }
...